requirements.txt
#swagger-client==1.0.0
pandas
numpy
requests
python-dotenv
cryptography
streamlit
matplotlib
seaborn
typing